+++ ChangeLog.md
@@ -3,6 +3,16 @@
 - *Geometry2d*
   - LUTBasedNSDistanceTransform: Removed the dependency to libNetPBM. (Nicolas Normand)
 
+# DGtalTools-contrib  0.9.3
+
+- *visualisation*
+  - displayLineSegments: add a new tool allowing to display line segment in an output image. (Bertrand Kerautret)
+  - meshViewerEdit: add a button to invert the current selection (Bertrand Kerautret)
+  
+- *Geometry2d*
+  - houghLineDetect: to detect line segment from Hough transform (using OpenCV). 
+  
+
 # DGtalTools-contrib  0.9.2
 
 - *visualisation*